refactor: extract BracketMatcher utility, fix dead code, add error logging
